Remove dead code an move remaining to a better place
